requirements.txt
pandas
numpy
matplotlib 
scikit-learn
torch
transformers
flask
deep_translator
tweepy
python-dotenv
gunicorn 